[addync_sync_to_suggestions] Adding a default node system to be used as template
