+++ src/generator/suggestions/node08.js
@@ -1,8 +1,8 @@
 import { _ } from 'azk';
-import { UIProxy } from 'azk/cli/ui';
-import { example_system } from 'azk/generator/rules';
 
-export class Suggestion extends UIProxy {
+import { Suggestion as DefaultSuggestion } from 'azk/generator/suggestions/node_default';
+
+export class Suggestion extends DefaultSuggestion {
   constructor(...args) {
     super(...args);
 
@@ -13,18 +13,8 @@ export class Suggestion extends UIProxy {
     this.ruleNamesList = ['node08'];
 
     // Initial Azkfile.js suggestion
-    this.suggestion = _.extend({}, example_system, {
-      __type: "node.js",
+    this.suggestion = _.extend({}, this.suggestion, {
       image : { docker: "node:0.8" },
-      provision: [
-        "npm install"
-      ],
-      http: true,
-      scalable: { default: 2 },
-      command : "npm start",
-      envs    : {
-        NODE_ENV: "dev"
-      }
     });
   }
